Baseball Lineup Simulation
Simulation comparison of a team of MacGuires vs a team of Tony Gwynns. Hill Climbing approach to a mixed lineup to generate an optimal lineup of batters using a 2-swap neighborhood and comparing the distributions of runs by t-tests to determine if the neighboring solution is better with acceptance of p<0.05.